{"types":"C","runs":32}

+++ ConsoleSample.java
@@ -4,7 +4,7 @@ public class ConsoleSample extends ConsoleProgram {
 	public void run() {
 	int n = readInt("n :");
 	int m = readInt ("m :");
-	int average = ((n+m)/2);
+	double average = ((n+m)/2);
 	println (average);
 }
 }
